Webclient/server refactor
