+++ internal/server/http/webclient/assets/wasm_exec.js
