+++ internal/server/http/webclient/compile.zsh
@@ -1,0 +1,1 @@
+GOARCH=wasm GOOS=js go build -o assets/lib.wasm
