--- internal/server/http/webserver/compile.zsh
@@ -1,1 +1,0 @@
-GOARCH=wasm GOOS=js go build -o lib.wasm
