+++ internal/server/http/webserver/ttlmap.go
@@ -1,4 +1,4 @@
-package apiserver
+package webserver
 
 // Credit to https://stackoverflow.com/questions/25484122/map-with-ttl-option-in-go
